+++ src/main/kotlin/com/lykke/matching/engine/daos/TradeInfo.kt
@@ -1,5 +1,6 @@
 package com.lykke.matching.engine.daos
 
+import java.math.BigDecimal
 import java.util.Date
 
-data class TradeInfo(val assetPair: String, val isBuy: Boolean, val price: Double, val date: Date)
+data class TradeInfo(val assetPair: String, val isBuy: Boolean, val price: BigDecimal, val date: Date)
